fix(models): only drop event notifications opt in column if it exists

+++ packages/server/models/user/migrations/1729262784-remove-event-notifications-opt-in.js
@@ -1,7 +1,16 @@
-exports.up = knex => {
-  return knex.schema.table('users', table => {
-    table.dropColumn('eventNotificationsOptIn')
-  })
+exports.up = async knex => {
+  const hasOptInColumn = await knex.schema.hasColumn(
+    'users',
+    'eventNotificationsOptIn',
+  )
+
+  if (hasOptInColumn) {
+    return knex.schema.table('users', table => {
+      table.dropColumn('eventNotificationsOptIn')
+    })
+  }
+
+  return true
 }
 
 /**
